% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_functions.R
\name{adjustLayer_byImp}
\alias{adjustLayer_byImp}
\title{Adjust a depth interval by the depth of a possible restriction}
\usage{
adjustLayer_byImp(depths, imp_depth, sdepths)
}
\arguments{
\item{depths}{A numeric vector of length two. The upper and lower depth
bound that should be adjusted.}

\item{imp_depth}{A numeric value. The depth of a restriction.}

\item{sdepths}{A numeric vector. The lower depths of soil layers,
sorted by increasing depth in a consistent unit.}

\item{The}{adjusted values of \code{depths}.}
}
\description{
Adjust a depth interval by the depth of a possible restriction
}
\examples{
depths_bottom <- c(10, 30, 50, 100)

adjustLayer_byImp(c(15, 70), 60, depths_bottom)
adjustLayer_byImp(c(15, 70), 5, depths_bottom)
adjustLayer_byImp(c(15, 70), 150, depths_bottom)

}
