% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_calculate.R
\name{calc_BareSoilEvapCoefs}
\alias{calc_BareSoilEvapCoefs}
\title{Calculate potential bare-soil evaporation coefficients}
\usage{
calc_BareSoilEvapCoefs(
  layers_depth,
  sand,
  clay,
  depth_max_bs_evap_cm = 15,
  method_bad_soils = c("stop", "pass"),
  noSoilValue = 0,
  tolerance = sqrt(.Machine[["double.eps"]])
)
}
\arguments{
\item{layers_depth}{A numeric vector, matrix, or data.frame. Values describe
the lower soil layer depths in units of centimeters.}

\item{sand}{A numeric vector, matrix, or data.frame. Values are sand contents
in units of mass-percentage / 100.}

\item{clay}{A numeric vector, matrix, or data.frame. Values are clay contents
in units of mass-percentage / 100.}

\item{depth_max_bs_evap_cm}{A numeric value. The maximal soil depth in
centimeters from which bare-soil evaporation is potentially drawing
moisture.}

\item{method_bad_soils}{A character string. The method that determines how
the function behaves if inputs represent "bad" soils, e.g., soil properties
are missing or values out of range: \itemize{
  \item "stop" will stop the function with an error;
  \item "pass" will return \code{NA} for the sites with bad inputs.
}}

\item{noSoilValue}{A numeric value or `NA` (default 0). Value used to fill
in output if there are missing soil depth or soil properties.}

\item{tolerance}{A numeric value. Used to adjust check if soil property
values are less or equal to `1`.}
}
\value{
A numeric matrix with potential bare-soil evaporation coefficients
  where rows correspond to sites and columns to soil layers.
}
\description{
Soil texture influence based on re-analysis of data from Wythers et al. 1999.
Default of \code{depth_max_bs_evap} = 15 cm from Torres et al. 2010.
}
\section{Notes}{
 Rows of soil input arguments \code{layers_depth},
  \code{sand}, and \code{clay} correspond to sites and columns to soil
  layers. If \code{sand} and/or \code{clay} are vectors, then they are
  converted to 1-row matrices. If \code{layers_depth} is a vector, then it is
  converted to a matrix with as many sites/rows as \code{sand} and
  \code{clay} have. That is the code assumes identical soil layer depths for
  each site. All soil input arguments must have the same number of sites
  and of soil layers, i.e., identical matrix dimensions.
}

\section{Warning}{
 Influence of gravel is not accounted for.
}

\examples{
sw_soils1 <- data.frame(
  depth_cm = c(5, 10, 20, 30, 40, 60, 80, 100),
  sand_frac = c(0.51, 0.44, 0.35, 0.32, 0.31, 0.32, 0.57, 0.57),
  clay_frac = c(0.15, 0.26, 0.41, 0.45, 0.47, 0.47, 0.28, 0.28)
)

calc_BareSoilEvapCoefs(
  layers_depth = sw_soils1[, "depth_cm"],
  sand = sw_soils1[, "sand_frac"],
  clay = sw_soils1[, "clay_frac"]
)

sw_soils2 <- list(
  depth_cm = t(data.frame(
    site1 = c(5, 10, 20),
    site2 = c(5, 10, 15),
    site3 = c(5, 15, 30),
    site4 = c(5, 10, 15),
    site5 = c(20, 30, NA)
  )),
  sand = t(data.frame(
    site1 = c(0.50, 0.40, 0.30),
    site2 = c(0.25, 0.30, 0.35),
    site3 = c(1.00, 0.30, 0.00),
    site4 = c(1.25, 1.00, NA),
    site5 = c(0.25, 0.30, 0.35)
  )),
  clay = t(data.frame(
    site1 = c(0.20, 0.20, 0.25),
    site2 = c(0.15, 0.25, 0.25),
    site3 = c(0.00, 0.25, 1.00),
    site4 = c(0.15, 0.00, 1.00),
    site5 = c(0.15, 0.25, 0.25)
  ))
)

calc_BareSoilEvapCoefs(
  layers_depth = sw_soils2[["depth_cm"]],
  sand = sw_soils2[["sand"]],
  clay = sw_soils2[["clay"]],
  method_bad_soils = "pass",
  noSoilValue = NA
)

}
\references{
Torres EA, Calera A (2010) Bare soil evaporation under high
  evaporation demand: a proposed modification to the FAO-56 model.
  Hydrological Sciences Journal- Journal des Sciences Hydrologiques, 55,
  303-315.

Wythers K.R., Lauenroth W.K., Paruelo J.M. (1999) Bare-Soil
  Evaporation Under Semiarid Field Conditions. Soil Science Society of
  America Journal, 63, 1341-1349.
}
