% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_functions.R
\name{calc_weights_from_depths}
\alias{calc_weights_from_depths}
\title{Calculate the relative contribution of existing soil layers at a new depth}
\usage{
calc_weights_from_depths(il_new, target_cm, depths_cm)
}
\arguments{
\item{il_new}{A numeric value.}

\item{target_cm}{A numeric value.}

\item{depths_cm}{A numeric vector.}
}
\value{
A numeric vector of length two.
}
\description{
Calculate the relative contribution of existing soil layers at a new depth
}
\examples{
target <- 45
depths <- c(10, 30, 50, 100)
calc_weights_from_depths(
  il_new = findInterval(target, depths),
  target_cm = target,
  depths_cm = depths
)

}
