% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SDA.R
\name{calculate_NRCS_soil_depth}
\alias{calculate_NRCS_soil_depth}
\title{Determine soil depth}
\usage{
calculate_NRCS_soil_depth(x, target_cokeys, restrict_by_ec_or_ph = TRUE)
}
\arguments{
\item{x}{A \code{data.frame}. The result of
function \code{\link{fetch_soils_from_NRCS_SDA}} that
returned horizon-level data for a set of \var{mukey} values.}

\item{target_cokeys}{A vector. The set of \var{cokey} values for which
soil depth is to be determined based on \code{x}.}

\item{restrict_by_ec_or_ph}{A logical value. Include depth restrictions
by \code{ph <= 3.5} or \code{ec >= 16}.}
}
\value{
A \code{data.frame} with at least three columns. Each row represents
  one value of \code{target_cokeys}. The columns are: \itemize{
    \item \var{N_horizons}: The number of soil horizons/layers.
    \item \var{SoilDepth_cm}: The soil depth in centimeters.
    \item \var{depth_Lx}: The lower depth of soil horizon/layer \var{x}.
      Note: \var{depth_L1} may vary from \code{x[, "hzdepb_r"]}.
  }
}
\description{
Determine soil depth
}
\examples{
\dontrun{
if (curl::has_internet()) {
  x <- fetch_soils_from_NRCS_SDA(mukeys_unique = c(471168, 1606800))

  calculate_NRCS_soil_depth(x, restrict_by_ec_or_ph = FALSE)

  x2 <- cbind(x, organic = is_NRCS_horizon_organic(x))
  calculate_NRCS_soil_depth(x2, restrict_by_ec_or_ph = TRUE)
}
}

}
\references{
Code based on \var{CalcRZDepth()} version 2020-08-31:
  \url{https://github.com/ncss-tech/SoilDataDevelopmentToolbox/blob/master/SDA_Valu2Table.py}
  Note: currently ignores "dense" layer restrictions
}
