% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_check.R
\name{check_depth_table}
\alias{check_depth_table}
\title{Check validity of a table with soil layer depths}
\usage{
check_depth_table(table_depths, soil_depth, n_layers)
}
\arguments{
\item{table_depths}{A two-dimensional numeric matrix-like object.
Rows represent sites and columns soil layers \var{1, ..., n}. Values
represent the lower depth of a layer.}

\item{soil_depth}{A numeric vector. The soil depth at each site in the same
units as \code{table_depths}. If missing, then calculated and not
checked.}

\item{n_layers}{A numeric vector. The number of soil layers at each site.
If missing, then calculated and not checked.}
}
\value{
If all checks succeed, then \code{TRUE}; otherwise, a named list with
  check results.
}
\description{
The function runs four checks: \itemize{
  \item \code{soil_depth} agrees with the deepest soil layer.
  \item \code{n_layers} agrees with the number of soil layers.
  \item Soil layers are organized continuously and begin at the surface.
  \item Depths of soil layers are strictly monotonic increasing.
  \item Each site has at least one soil layer.
}
}
\examples{
soils <- matrix(
  data = c(
    5, 10, 50, 150,
    5, 10, 50, 200,
    5, 10, 50, 150,
    5, NA, 50, 150,
    5, 5, 50, 150,
    NA, NA, NA, NA
  ),
  nrow = 6,
  byrow = TRUE
)
sd_cm <- c(150, 150, 150, 150, 150, 0)
nlyrs <- c(4, 4, 3, 3, 4, 0)

# This illustrates the warnings
soil_checks1 <- suppressWarnings(check_depth_table(
  table_depths = soils,
  soil_depth = sd_cm,
  n_layers = nlyrs
))

# This should be ok
soil_checks2 <- check_depth_table(
  table_depths = soils[1, , drop = FALSE],
  soil_depth = sd_cm[1],
  n_layers = nlyrs[1]
)

stopifnot(soil_checks2)

}
