% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_POLARIS.R
\name{check_POLARIS}
\alias{check_POLARIS}
\title{Check that \var{POLARIS} soil data are locally available}
\usage{
check_POLARIS(
  path = ".",
  vars = c("bd", "sand", "clay", "silt"),
  stats = "mean"
)
}
\arguments{
\item{path}{A character string. The path to the local copy of the
\var{POLARIS} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_POLARIS())}.}

\item{vars}{A vector of character strings. See Chaney et al. 2019}

\item{stats}{A vector of character strings. See Chaney et al. 2019}
}
\value{
A logical array with four dimensions: \describe{
  \item{File type}{
    \var{"vrt"} or \var{"tif"}; if \var{"vrt"} is missing, then \var{"tif"}
    are not checked.}
  \item{Variables}{Checks for each \code{vars}.}
  \item{Statistics}{Checks for each \code{stats}.}
  \item{Soil layer depths}{Checks for each soil layer in \var{POLARIS}.}
}
}
\description{
Check that \var{POLARIS} soil data are locally available
}
\examples{
script_to_download_polaris <- prepare_script_for_POLARIS()

## Execute script to download data
## (or set `path_polaris` to your local copy)

path_polaris <- dirname(script_to_download_polaris)
vars <- c("bd", "sand", "clay", "silt")
stat <- "mean"

## Check that we have POLARIS data
has_POLARIS <- check_POLARIS(path = path_polaris, vars = vars, stat = stat)

# Do we have all files?
isTRUE(all(has_POLARIS))

# If not, then examine
# (i) whether vrt files are missing, and/or
has_vrt <- !is.na(has_POLARIS["vrt", , , ]) & has_POLARIS["vrt", , , ]

# (ii) whether tif files are missing
has_tif <- !is.na(has_POLARIS["tif", , , ]) & has_POLARIS["tif", , , ]

}
\references{
Chaney, N. W., B. Minasny, J. D. Herman, T. W. Nauman, C. Brungard,
 C. L. S. Morgan, A. B. McBratney, E. F. Wood, and Y. T. Yimam. 2019.
 POLARIS soil properties: 30-meter probabilistic maps of soil properties
 over the contiguous United States. Water Resources Research 55:2916-2938.
 \url{https://doi.org/10.1029/2018WR022797}.
}
