% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_calculate.R
\name{deduce_complete_soil_texture}
\alias{deduce_complete_soil_texture}
\title{Deduce soil texture \var{iff} one of sand, clay, silt is missing}
\usage{
deduce_complete_soil_texture(
  x,
  var_stxt = c("sand", "clay", "silt"),
  val_total = 1,
  ignore_le = 0
)
}
\arguments{
\item{x}{A \code{data.frame}. Soil horizons/layers are organized in rows
and soil texture variables in columns.}

\item{var_stxt}{A vector of character strings. The names of the
three columns representing sand, clay, and silt.}

\item{val_total}{A numerical value. The total value that the sum of
sand, clay, and silt should be.}

\item{ignore_le}{A numerical value. Single missing values are not deduced
if the two other values sum to less or equal to \code{ignore_le}.}
}
\description{
Deduce soil texture \var{iff} one of sand, clay, silt is missing
}
\examples{
x <- data.frame(
  sand_pct = c(45.7, NA, 68.5, NA, 0, 2, NA),
  clay_pct = c(12.5, NA, NA, NA, 0, NA, 10),
  silt_pct = c(41.8, NA, 21.5, 15, NA, 1, 15)
)

x2 <- deduce_complete_soil_texture(
  x = x,
  var_stxt = c("sand_pct", "clay_pct", "silt_pct"),
  val_total = 100,
  ignore_le = 5
)
x2[3, ] # previously missing clay was deduced to 10\%
x2[5, ] # missing silt was not filled in because others sum to <= 5
x2[6, ] # missing clay was not filled in because others sum to <= 5
x2[7, ] # previously missing sand was deduced to 75\%

}
