% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_calculate.R
\name{estimate_bulkdensity}
\alias{estimate_bulkdensity}
\title{Estimate soil bulk density}
\usage{
estimate_bulkdensity(theta_saturated, gravel_volume)
}
\arguments{
\item{theta_saturated}{A numeric vector. Saturated volumetric water
content, i.e., at 0 \var{MPa}, in units of \var{cm3 / cm3}.}

\item{gravel_volume}{A numeric vector. Volume fraction of gravel/rock
in units of \var{cm3 / cm3}}
}
\value{
Bulk density in units of \var{g / cm3}
}
\description{
Estimate soil bulk density
}
\references{
Saxton, K. E., and W. J. Rawls. 2006.
Soil water characteristic estimates by texture and organic matter for
hydrologic solutions. Soil Science Society of America Journal 70:1569–1578.
}
