% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SDA.R
\name{extract_soils_NRCS_SDA}
\alias{extract_soils_NRCS_SDA}
\title{Extract soil information from the Soil Data Access \var{SDA} service by
\var{NRCS} for \pkg{SOILWAT2} applications}
\usage{
extract_soils_NRCS_SDA(
  x,
  crs = 4326,
  mukeys,
  method = c("SSURGO", "STATSGO", "SSURGO_then_STATSGO"),
  remove_organic_horizons = c("none", "all", "at_surface"),
  replace_missing_fragvol_with_zero = c("none", "all", "at_surface"),
  estimate_missing_bulkdensity = FALSE,
  restrict_by_ec_or_ph = TRUE,
  impute_locf = FALSE,
  digits = 3L,
  chunk_size = 1000L,
  progress_bar = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-classes]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{mukeys}{A character or integer vector. List of soil map unit keys
for which soil information should be extracted. Provide \code{locations}
or \code{mukeys}.}

\item{method}{A character string. Method indicating whether \var{SDA}
should query \var{"SSURGO"}, \var{"STATSGO"}, or
\var{"SSURGO_then_STATSGO"} which attempts to replace
\code{locations} without \var{"SSURGO"} soil information,
e.g., due to unmapped areas, with \var{"STATSGO"} data.}

\item{remove_organic_horizons}{A character string. Method
indicating how to deal with organic horizons as determined by
function \code{\link{is_NRCS_horizon_organic}}. See details.}

\item{replace_missing_fragvol_with_zero}{A character string. Method
indicating how missing/null values of rock/gravel fragment fractions
should be interpreted. See details.}

\item{estimate_missing_bulkdensity}{A logical value. Estimate missing
bulk density values from saturated water content and gravel volume.
See \code{\link{estimate_bulkdensity}}.}

\item{restrict_by_ec_or_ph}{A logical value. Include depth restrictions
by \code{ph <= 3.5} or \code{ec >= 16}.}

\item{impute_locf}{A logical value. Impute missing values with a
shallow-depth value carried deeper approach (in analogy to \var{LOCF}).
Consequently, missing values in the shallowest horizon are not imputed.
See \code{\link[rSW2utils]{impute_df}}.}

\item{digits}{An integer value. The number of digits to which soil texture
variables are rounded. Skip rounding if \code{NA} or \code{NULL}.}

\item{chunk_size}{An integer value. The size of chunks into which
\code{locations} is broken up and looped over for processing.}

\item{progress_bar}{A logical value. Display a progress bar as the code
loops over the chunks?}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the Soil Data Access \var{SDA} service by
\var{NRCS} for \pkg{SOILWAT2} applications
}
\section{Details}{
 \var{NRCS} soil datasets \var{SSURGO} and \var{STATSGO} are
  organized in soil map units \var{mukey} that are
  spatially explicit (i.e., we can query their values by geographic location)
  and within each \var{mukey} into soil map unit components \var{cokey}
  which have no explicit spatial arrangement within a soil map unit.
  Because soil texture information is specific to soil map unit components,
  geographic location alone is insufficient to query soil texture.

  This function relies that soil information of exactly one \var{cokey}
  per each \code{location} or \code{mukeys} is returned by
  \code{\link{fetch_soils_from_NRCS_SDA}}. The default \var{SQL} template
  \var{"nrcs_sql_template.sql"} extracts the "dominant component".
  The dominant component is defined as the the first \var{cokey} with the
  highest representative component percent \var{comppct_r}.
  See \var{GetDominantComponent.py}
  from \url{https://github.com/ncss-tech/SoilDataDevelopmentToolbox}.


  The argument \code{remove_organic_horizons} is one of
  \describe{
    \item{\var{"all"}}{
      All organic layers (at surface or buried) are removed and horizon
      number and depths are recalculated.
    }
    \item{\var{"at_surface"}}{
      Organic layer(s) at the soil surface are removed and horizon
      number and depths are recalculated. Buried organic horizons remain
      unmodified.
    }
    \item{\var{"none"}}{
       Horizons are not modified.
    }
  }


  The argument \code{replace_missing_fragvol_with_zero} is one of
  \describe{
    \item{\var{"all"}}{
      Missing/null values of rack/gravel fragments are
      replaced by 0 %.
      See also argument \var{nullFragsAreZero} of
      function \code{\link[soilDB]{fetchSDA}}.
    }
    \item{\var{"at_surface"}}{
      Missing/null values of rack/gravel fragments are
      replaced by 0 % in the shallowest horizon only.
      Note, remaining missing values in deeper horizons
      can subsequently be imputed by argument \code{impute_locf}.
    }
    \item{\var{"none"}}{
       Missing/null values remain unmodified unless
       argument \code{impute_locf} is activated.
    }
  }
}

\section{Notes}{
 A live internet connection is required to access \var{SDA}.
}

\examples{
\dontrun{
if (curl::has_internet()) {
  locations <- matrix(
    data = c(-120.325, -111.245, 39.855, 36.753),
    nrow = 2
  )

  # Example 1: extract soils by mukey values
  extract_soils_NRCS_SDA(mukeys = c(471168, 1606800))

  # Example 2: extract soils by geographic location
  extract_soils_NRCS_SDA(x = locations)

  # Example 3: first identify mukey values by geographic location,
  # then query soils from SSURGO by mukey,
  # but still pass locations in case we need to query STATSGO as well

  mukeys <- fetch_mukeys_spatially_NRCS_SDA(
    x = locations,
    db = "SSURGO",
    progress_bar = TRUE
  )

  extract_soils_NRCS_SDA(
    x = locations,
    mukeys = mukeys[["mukeys"]],
    method = "SSURGO_then_STATSGO",
    remove_organic_horizons = "at_surface",
    replace_missing_fragvol_with_zero = "at_surface",
    estimate_missing_bulkdensity = TRUE,
    restrict_by_ec_or_ph = FALSE,
    impute_locf = TRUE,
    progress_bar = TRUE,
    verbose = TRUE
  )
}
}

}
\seealso{
\code{\link[soilDB]{fetchSDA}} and \code{\link[soilDB]{SDA_query}}
}
