% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_POLARIS.R
\name{extract_soils_POLARIS}
\alias{extract_soils_POLARIS}
\title{Extract soil information from the \var{POLARIS} soil dataset
for \pkg{SOILWAT2} applications}
\usage{
extract_soils_POLARIS(
  x,
  crs = 4326,
  vars = c("bd", "sand", "clay", "silt"),
  stat = "mean",
  path,
  method = c("asis", "fix_with_buffer"),
  fix_criteria = list(bd = list(op = "<", value = 0.6), texture = list(op = "<", value
    = 0.5)),
  buffer_m = NULL,
  fun = NULL,
  na.rm = TRUE,
  digits = 3L,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-classes]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{vars}{A vector of character strings. See Chaney et al. 2019}

\item{stat}{A character string. See Chaney et al. 2019}

\item{path}{A character string. The path to the local copy of the
\var{POLARIS} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_POLARIS())}.}

\item{method}{A character string. Method that determines extraction approach:
(i) values are extracted using arguments
    \code{buffer_m}, \code{fun}, and \code{na.rm}
    and are returned \var{"asis"} or
(ii) values are extracted for point locations,
    i.e., temporarily setting \code{buffer_m = NULL}; then,
    sites with problematic values (as determined by \code{fix_criteria})
    are extracted again under \var{"fix_with_buffer"} based on
    \code{buffer_m}, \code{fun}, and \code{na.rm}}

\item{fix_criteria}{A named list. Names correspond to \code{vars} or
to \var{"texture"} if criterion is to be applied to the sum of
sand, clay, and silt. Each element is applied to the variable of the
element name to determine whether a site has problematic values.
Elements are each a named list with two elements
\var{"op"} for the relationship operator, e.g., \var{"<"}, and
\var{"value"} for the value to compare against. See examples.}

\item{buffer_m}{A numeric value. The radius of a buffer around each point
from which to extract cell values and across which \code{fun} is applied.
Passed to \code{\link[raster]{extract}}. Set to \code{NULL} to extract
\var{POLARIS} gridcell values at point locations.}

\item{fun}{A function if \code{method} is either value or
a named list of functions if \code{method = "fix_with_buffer"} where
names correspond to \code{vars} or to \var{"texture"} if function is to
be applied (individually) to sand, clay, and silt.
Summarizing gridcell values if more than one value
is extracted per location. See \code{\link[raster]{extract}}.}

\item{na.rm}{A logical value. Passed to \code{fun}.}

\item{digits}{An integer value. The number of digits to which soil texture
variables are rounded. Skip rounding if \code{NA} or \code{NULL}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \var{POLARIS} soil dataset
for \pkg{SOILWAT2} applications
}
\section{Notes}{
 A local copy of \var{POLARIS} is required. The function
  \code{\link{prepare_script_for_POLARIS}} creates a script that can be used
  to download \var{POLARIS} files.

 \var{POLARIS} uses weight-based percent as unit for
  \var{sand}, \var{clay}, \var{silt}; values occur in
  1% increments within \code{[0.5, 98.5]%}. However, the function returns
  soil texture in units of weight-based fractions.
}

\examples{
script_to_download_polaris <- prepare_script_for_POLARIS()

## Execute script to download data
## (or set `path_polaris` to your local copy)

path_polaris <- dirname(script_to_download_polaris)
vars <- c("bd", "sand", "clay", "silt")
stat <- "mean"

## Check that we have POLARIS data
has_POLARIS <- isTRUE(all(
  check_POLARIS(path = path_polaris, vars = vars, stat = stat)
))

if (has_POLARIS) {

  locations <- matrix(
    data = c(-120.1286878, -111.8511136, 39.8182913, 36.9047396),
    nrow = 2
  )

  ## Extract median of mean gridcell values across 100-m buffer
  ## around point locations
  res1 <- extract_soils_POLARIS(
    x = locations,
    vars = vars,
    stat = stat,
    path = path_polaris,
    buffer_m = 100,
    fun = median,
    na.rm = TRUE
  )

  ## Extract mean gridcell values at point locations and use 70-m buffer at
  ## sites with bad values
  res2 <- extract_soils_POLARIS(
    x = locations,
    vars = vars,
    stat = stat,
    path = path_polaris,
    method = "fix_with_buffer",
    fix_criteria = list(
      bd = list(op = "<", value = 0.6),
      texture = list(op = "<", value = 50)
    ),
    buffer_m = 70,
    fun = list(
      bd = function(x, na.rm = TRUE) median(x[x > 0.6], na.rm = na.rm),
      texture = median
    ),
    na.rm = TRUE,
    digits = 3
  )
}

# Clean up example
unlink(script_to_download_polaris)


}
\references{
Chaney, N. W., B. Minasny, J. D. Herman, T. W. Nauman, C. Brungard,
 C. L. S. Morgan, A. B. McBratney, E. F. Wood, and Y. T. Yimam. 2019.
 POLARIS soil properties: 30-meter probabilistic maps of soil properties
 over the contiguous United States. Water Resources Research 55:2916-2938.
 \url{https://doi.org/10.1029/2018WR022797}.
}
\seealso{
\code{\link[raster]{extract}}
}
