% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SDA.R
\name{fetch_mukeys_spatially_NRCS_SDA}
\alias{fetch_mukeys_spatially_NRCS_SDA}
\title{Spatially query \var{mukey} values for point locations from \var{NRCS}
web-based \var{SDA} service}
\usage{
fetch_mukeys_spatially_NRCS_SDA(
  x,
  crs = 4326,
  db = c("SSURGO", "STATSGO"),
  ...,
  chunk_size = 50L,
  progress_bar = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-classes]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{db}{A character string. Query \var{mukey} from the \var{SSURGO} or
from the \var{STATSGO} soil database.}

\item{...}{Currently ignored.}

\item{chunk_size}{An integer value. The size of chunks into which
\code{locations} is broken up and looped over for processing.}

\item{progress_bar}{A logical value. Display a progress bar as the code
loops over the chunks?}
}
\value{
A named list with two elements: \itemize{
  \item{\var{ref}} The data reference.
  \item{\var{mukeys}} A vector with a \var{mukey} value for each
    \code{locations}.
}
}
\description{
Spatially query \var{mukey} values for point locations from \var{NRCS}
web-based \var{SDA} service
}
\examples{
\dontrun{
if (curl::has_internet()) {
  locations <- matrix(
    data = c(-120.325, -111.245, 39.855, 36.753),
    nrow = 2
  )

  fetch_mukeys_spatially_NRCS_SDA(locations)
}
}

}
