% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_POLARIS.R
\name{fetch_soils_from_POLARIS}
\alias{fetch_soils_from_POLARIS}
\title{Extract soil information from the \var{POLARIS} soil dataset}
\usage{
fetch_soils_from_POLARIS(
  x,
  crs,
  vars,
  stat,
  path,
  buffer_m = NULL,
  fun = NULL,
  na.rm = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-classes]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{vars}{A vector of character strings. See Chaney et al. 2019}

\item{stat}{A character string. See Chaney et al. 2019}

\item{path}{A character string. The path to the local copy of the
\var{POLARIS} folder hierarchy, e.g.,
\code{dirname(prepare_script_for_POLARIS())}.}

\item{buffer_m}{A numeric value. The radius of a buffer around each point
from which to extract cell values and across which \code{fun} is applied.
Passed to \code{\link[raster]{extract}}. Set to \code{NULL} to extract
\var{POLARIS} gridcell values at point locations.}

\item{fun}{A function. Summarizing gridcell values if more than one value
is extracted per location. See \code{\link[raster]{extract}}.}

\item{na.rm}{A logical value. Passed to \code{fun}.}

\item{verbose}{A logical value.}
}
\description{
Extract soil information from the \var{POLARIS} soil dataset
}
\section{Notes}{
 This is a function with minimal functionality;
use \code{\link{extract_soils_POLARIS}} for a user-friendly interface.
}

\references{
Chaney, N. W., B. Minasny, J. D. Herman, T. W. Nauman, C. Brungard,
 C. L. S. Morgan, A. B. McBratney, E. F. Wood, and Y. T. Yimam. 2019.
 POLARIS soil properties: 30-meter probabilistic maps of soil properties
 over the contiguous United States. Water Resources Research 55:2916-2938.
 \url{https://doi.org/10.1029/2018WR022797}.
}
