% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_functions.R
\name{getLayersWidth}
\alias{getLayersWidth}
\title{Determine soil layer widths from depth profile}
\usage{
getLayersWidth(layers_depth)
}
\arguments{
\item{layers_depth}{A numeric vector. The lower depths of soil layers,
sorted by increasing depth in a consistent unit.}
}
\description{
Determine soil layer widths from depth profile
}
\examples{
depths_bottom <- c(10, 30, 50, 100)
widths <- getLayersWidth(depths)

data.frame(
  layer_depth_top = depths_bottom - widths,
  layer_width = widths,
  layer_depth_bottom = depths_bottom
)

}
