% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_functions.R
\name{identify_soillayers}
\alias{identify_soillayers}
\title{Find the soil layer numbers in a soil layer profile that are affected by
a soil depth or a soil band.}
\usage{
identify_soillayers(depths, sdepth)
}
\arguments{
\item{depths}{A numeric vector of length one or two.}

\item{sdepth}{A numeric vector. The lower depths of soil layers,
sorted by increasing depth in a consistent unit.}
}
\description{
Find the soil layer numbers in a soil layer profile that are affected by
a soil depth or a soil band.
}
\examples{
depths_bottom <- c(10, 30, 50, 100)
identify_soillayers(5, depths_bottom)
identify_soillayers(c(15, 40), depths_bottom)
identify_soillayers(c(15, 150), depths_bottom)

}
