% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_calculate.R
\name{impute_soils}
\alias{impute_soils}
\title{Impute missing soil values per location}
\usage{
impute_soils(
  x,
  var_values,
  var_site_id = "COKEY",
  var_horizon = "Horizon_No",
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.
Soil horizons/layers are organized in rows
and soil texture variables in columns.}

\item{var_values}{A vector of character strings. The column names of \code{x}
which contain the soil variables to be imputed.}

\item{var_site_id}{A character string. The column name of \code{x} which
contains the unique location identifiers.}

\item{var_horizon}{A character string. The column name of \code{x} which
contains the horizon/layer numbers where the shallowest horizon/layer
is number one.}

\item{verbose}{A logical value.}
}
\value{
An updated version of \code{x}.
}
\description{
Impute missing soil values per location by shallow-depth value carried
deeper (in analogy to \var{LOCF}), but do not impute missing values
in the shallowest horizon/layer.
}
\examples{
x <- data.frame(
  id = rep(1, 7),
  layer_no = 1:7,
  coarse = c(NA, 10, 50, NA, 0, 15, NA),
  sand_pct = c(45.7, NA, 68.5, NA, 0, 2, NA),
  clay_pct = c(12.5, NA, NA, NA, 0, NA, 10),
  silt_pct = c(41.8, NA, 21.5, 15, NA, 1, 15)
)

# Missing values in the shallowest layer are not imputed
impute_soils(
  x,
  var_values = c("coarse", "sand_pct"),
  var_site_id = "id",
  var_horizon = "layer_no"
)

}
\seealso{
\code{\link[rSW2utils]{impute_df}}
}
