% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_calculate.R
\name{init_soiltemperature}
\alias{init_soiltemperature}
\title{Crude initial estimate of soil temperatures throughout the soil profile}
\usage{
init_soiltemperature(
  layers_depth,
  Tsoil_upper,
  Tsoil_const,
  depth_Tsoil_const = 990
)
}
\arguments{
\item{layers_depth}{A numeric vector. Values describe
the lower soil layer depths in units of centimeters at which soil
temperatures are estimated.}

\item{Tsoil_upper}{A numeric value. Initial soil temperature [C] at the
upper boundary, i.e., soil surface.}

\item{Tsoil_const}{A numeric value. Initial soil temperature [C] at the
lower boundary, i.e., the soil depth at which soil temperature is mostly
constant throughout a year.}

\item{depth_Tsoil_const}{A numeric value. The soil depth [cm] for which
\code{Tsoil_const} is valid.}
}
\value{
A numeric vector of length corresponding to \code{layers_depth}
}
\description{
Crude initial estimate of soil temperatures throughout the soil profile
}
\examples{
init_soiltemperature(
  layers_depth = c(5, 10, 20, 30, 40, 60, 80, 100),
  # soil surface underneath a snow pack on Jan 1
  Tsoil_upper = -1,
  # mean annual air temperature is a reasonable estimate
  Tsoil_const = 8,
  depth_Tsoil_const = 990
)

}
