% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_NRCS_SDA.R
\name{is_NRCS_horizon_organic}
\alias{is_NRCS_horizon_organic}
\title{Check whether a \var{NRCS} soil horizon is organic}
\usage{
is_NRCS_horizon_organic(x)
}
\arguments{
\item{x}{A two-dimensional character object. Required columns are
\var{taxorder}, \var{taxsubgrp}, \var{desgnmaster}, \var{texture},
and \var{lieutex}. Each horizon is represented by exactly one row.}
}
\value{
A logical vector representing each horizon. \code{TRUE} indicates
  that a horizon is considered organic; default is \code{FALSE}
  representing mineral soils. \code{NA}s in the input propagate.
}
\description{
Based on function \code{CheckTexture()}, version 2020-08-31, of the
\var{SoilDataDevelopmentToolbox}.
}
\examples{
x <- data.frame(
  taxorder = c("Histosols", "x", "x", "x", "x", "x", NA),
  taxsubgrp = c("x", "histic", "x", "x", "x", "x", NA),
  desgnmaster = c("L", "L", "O", "x", "x", "x", NA),
  texture = c("x", "x", "x", "CE", "x", "x", NA),
  lieutex = c("x", "x", "x", "x", "Muck", "x", NA)
)

cbind(
  organic = is_NRCS_horizon_organic(x),
  expected_organic = c(FALSE, FALSE, TRUE, TRUE, TRUE, FALSE, NA)
)

\dontrun{
if (curl::has_internet()) {
  x <- fetch_soils_from_NRCS_SDA(mukeys_unique = c(471168, 1606800))
  is_NRCS_horizon_organic(x)
}
}

}
\references{
Code based on \var{CheckTexture()} version \var{2020-Aug-31} from
  \url{https://github.com/ncss-tech/SoilDataDevelopmentToolbox/blob/master/SDA_Valu2Table.py}
}
