% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_soils_POLARIS.R
\name{prepare_script_for_POLARIS}
\alias{prepare_script_for_POLARIS}
\title{Create \var{wget} script to download soil data files from \var{POLARIS}}
\usage{
prepare_script_for_POLARIS(
  path = ".",
  version = "v1.0",
  vars = c("bd", "sand", "clay", "silt"),
  stat = "mean"
)
}
\arguments{
\item{path}{A character string. The path to where the local copy of the
\var{POLARIS} folder hierarchy and files should be downloaded.}

\item{version}{A character string. The \var{POLARIS} release version.}

\item{vars}{A vector of character strings. See Chaney et al. 2019}

\item{stat}{A vector of character strings. See Chaney et al. 2019}
}
\value{
(Invisibly) the file path of the generated bash script.
}
\description{
This function only writes out a text file to disk. The user
is responsible to run the script after making sure that \var{wget} is
available and the script is executable.
}
\examples{
fname_wget_polaris <- prepare_script_for_POLARIS()

## in a shell
## give execute permission if needed: chmod +x <fname_wget_polaris>
## download data: ./<fname_wget_polaris>

unlink(fname_wget_polaris)

}
\references{
Chaney, N. W., B. Minasny, J. D. Herman, T. W. Nauman, C. Brungard,
 C. L. S. Morgan, A. B. McBratney, E. F. Wood, and Y. T. Yimam. 2019.
 POLARIS soil properties: 30-meter probabilistic maps of soil properties
 over the contiguous United States. Water Resources Research 55:2916-2938.
 \url{https://doi.org/10.1029/2018WR022797}.
}
