% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_functions.R
\name{reshape_soilproperties}
\alias{reshape_soilproperties}
\alias{reshape_soilproperties_to_long}
\alias{reshape_soilproperties_to_wide}
\title{Reshape soil properties between wide format and (semi-)long format}
\usage{
reshape_soilproperties_to_long(
  x_wide,
  type_to = c("long", "long_by_properties"),
  id_site = "site",
  id_soillayer = "soillayer",
  sep_soillayer = "_L",
  soilproperties = c("db", "fsand", "fclay", "fcoarse"),
  id_values = "value",
  id_property = "variable"
)

reshape_soilproperties_to_wide(
  x_long,
  type_from = c("long", "long_by_properties"),
  id_site = "site",
  id_soillayer = "soillayer",
  sep_soillayer = "_L",
  soilproperties = c("db", "fsand", "fclay", "fcoarse"),
  id_values = "value",
  id_property = "variable"
)
}
\arguments{
\item{x_wide}{A data frame in wide format (see details).}

\item{type_to}{A character string. Determines if output format
is long or semi-long (see details).}

\item{id_site}{A character string. Column name of \code{x} that
contains unique site names.}

\item{id_soillayer}{A character string. Name of the column that will
be created to contain soil layer numbers.}

\item{sep_soillayer}{A character string. See details.}

\item{soilproperties}{A character vector. Names of soil properties;
see details.}

\item{id_values}{A character string. Name of the column in long format
that contains values of soil properties.}

\item{id_property}{A character string. Name of the column in long format
that contains names of soil properties.}

\item{x_long}{A data frame in long or semi-long format (see details).}

\item{type_from}{A character string. Describes if format of \code{x_long} is
long or semi-long.}
}
\value{
\itemize{
\item \code{\link[=reshape_soilproperties_to_long]{reshape_soilproperties_to_long()}} returns
\itemize{
\item a data frame in long format if \code{type_to} is \code{"long"}
\item a data frame in semi-long format if \code{type_to} is \code{"long_by_properties"}.
}
\item \code{\link[=reshape_soilproperties_to_long]{reshape_soilproperties_to_long()}} returns a data frame in wide format.
}
}
\description{
Reshape soil properties between wide format and (semi-)long format
}
\section{Details}{

Formats of soil-property data frames
\itemize{
\item \code{"wide"}
\itemize{
\item each row represents a unique site
(encoded by column \code{id_site})
\item besides \code{id_site}, columns represent soil properties
at different soil layers and column names follow the pattern
\verb{<soilproperties><sep_soillayer><soil layer id>}, e.g.,
\code{db_L1}, \code{fsand_L1}, \code{fclay_L1}, \code{fcoarse_L1}, \code{db_L2}, \code{fsand_L2}, ...
}
\item \code{"semi-long"} (or \code{"long_by_properties"})
\itemize{
\item each row represents a unique combination of sites and soil layers
(encoded by columns \code{id_site} and \code{id_soillayer})
\item columns are \code{id_site}, \code{id_soillayer}, \code{soilproperties}, e.g.,
\code{site}, \code{soillayer}, \code{db}, \code{fsand}, \code{fclay}, \code{fcoarse}
}
\item \code{"long"}
\itemize{
\item each row represents a unique observation, e.g.,
\code{SiteA}, 1, \code{fsand}, 0.70
\item columns are \code{id_site}, \code{id_soillayer}, \code{id_property}, \code{id_values}
}
}
}

\examples{
x_wide <- data.frame(
  location = c("SiteA", "SiteB"),
  db_L1 = c(1.5, 1.6),
  fsand_L1 = c(0.7, 0.2),
  fclay_L1 = c(0.1, 0.2),
  db_L2 = c(1.6, 1.7),
  fsand_L2 = c(0.75, 0.3),
  fclay_L2 = c(0.1, 0.15)
)

x_long <- reshape_soilproperties_to_long(
  x_wide,
  type_to = "long",
  id_site = "location",
  soilproperties = c("db", "fsand", "fclay")
)
x_wide_from_long <- reshape_soilproperties_to_wide(
  x_long,
  type_from = "long",
  id_site = "location",
  soilproperties = c("db", "fsand", "fclay")
)
all.equal(x_wide_from_long, x_wide, check.attributes = FALSE)

x_semilong <- reshape_soilproperties_to_long(
  x_wide,
  type_to = "long_by_properties",
  id_site = "location",
  soilproperties = c("db", "fsand", "fclay")
)
x_wide_from_semilong <- reshape_soilproperties_to_wide(
  x_semilong,
  type_from = "long_by_properties",
  id_site = "location",
  soilproperties = c("db", "fsand", "fclay")
)
all.equal(x_wide_from_semilong, x_wide, check.attributes = FALSE)

}
\seealso{
\code{\link[stats:reshape]{stats::reshape()}}
}
