% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2data-defunct.R
\name{rSW2data-defunct}
\alias{rSW2data-defunct}
\alias{calc_weights_from_depths}
\alias{add_layer_to_soil}
\title{Defunct functions in package \pkg{rSW2data}}
\usage{
calc_weights_from_depths(...)

add_layer_to_soil(...)
}
\arguments{
\item{...}{Previous function arguments.}
}
\description{
Executing a defunct function will fail and tell you which function
replaces them.
}
