% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2data-package.R
\docType{package}
\name{rSW2data}
\alias{rSW2data}
\alias{rSW2data-package}
\title{Package \pkg{rSW2data}: summary information}
\description{
Collection of functions to prepare and calculate input data for
  rSOILWAT2, rSFSW2, and rSFSTEP2 simulation experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrylandEcology/rSW2data}
  \item Report bugs at \url{https://github.com/DrylandEcology/rSW2data/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

}
