% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_calculate.R
\name{set_missing_soils_to_value}
\alias{set_missing_soils_to_value}
\title{Replace missing values of a soil variables with a new value}
\usage{
set_missing_soils_to_value(
  x,
  variable,
  value = 0,
  where = c("none", "all", "at_surface"),
  horizon = colnames(x)[1],
  verbose = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} or \code{matrix}.
Soil horizons/layers are organized in rows and variables in columns.}

\item{variable}{A character string. The variable name, i.e.,
one of the column names of \code{x}.}

\item{value}{The replacement value for missing values in
\code{x[, variable]}.}

\item{where}{A character string. Method to specify how and where in the
soil profile missing values of a variable should be replaced.
Possible options are:
\describe{
  \item{\var{"all"}}{
    Missing values are replaced by \code{value}.
  }
  \item{\var{"at_surface"}}{
    Missing values in the shallowest soil horizon
    are replaced by \code{value}.
    Note, remaining missing values in deeper horizons
    can subsequently be imputed, e.g.,
    \code{\link{impute_soils}}.
  }
  \item{\var{"none"}}{
     Missing values are set to \code{NA}.
  }
}}

\item{horizon}{A character string or integer vector. Either a column name
of \code{x} or a vector itself containing soil horizon/layer numbers that
start at 1 in the most shallow/surface horizon/layer.}

\item{verbose}{A logical value.}
}
\value{
An updated copy of \code{x}.
}
\description{
Replace missing values of a soil variables with a new value
}
\section{Details}{
 Missing values are those that are not finite, i.e.,
  not one of \code{NA}, \code{NaN}, or \code{Inf}.
}

\examples{
x <- data.frame(
  coarse = c(NaN, 10, 50, Inf, 0, 15, NA),
  sand_pct = c(45.7, NA, 68.5, NA, 0, 2, NA),
  clay_pct = c(12.5, NA, NA, NA, 0, NA, 10),
  silt_pct = c(41.8, NA, 21.5, 15, NA, 1, 15)
)

set_missing_soils_to_value(x, "coarse", where = "none")
set_missing_soils_to_value(x, "coarse", where = "all")
set_missing_soils_to_value(
  x = cbind(Layer_ID = seq_len(nrow(x)), x),
  variable = "coarse",
  where = "at_surface",
  horizon = "Layer_ID"
)
set_missing_soils_to_value(
  x,
  variable = "coarse",
  where = "at_surface",
  horizon = seq_len(nrow(x))
)

}
