% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soils_functions.R
\name{update_soil_profile}
\alias{update_soil_profile}
\title{Add/remove soil layers to soil data}
\usage{
update_soil_profile(
  soil_layers,
  requested_soil_layers,
  soil_data,
  variables = NULL,
  vars_exhaust = c("EvapCoeff", "TranspCoeff", "Imperm"),
  keep_prev_soildepth = TRUE,
  keep_prev_soillayers = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{soil_layers}{A two-dimensional numeric matrix or \code{data.frame}.
Rows represents sites; columns represent soil layers/horizons;
and values are depths of the lower limits of soil layers/horizons in
centimeters.}

\item{requested_soil_layers}{A numeric vector. The soil depths that should
be added if not already present.}

\item{soil_data}{A two-dimensional numeric matrix or \code{data.frame}.
Rows represents sites and columns represent
variables x soil layers/horizons combinations.
Column names are formatted as \var{var_Lx} where \var{x} represents
an increasing soil layer/horizon number.}

\item{variables}{A vector of character strings. The patterns identifying
all variables that are to be updated.}

\item{vars_exhaust}{A vector of character strings. The patterns
identifying variables whose values are "exhausted" when a soil layer is
split into two layers. See \code{\link{add_layer_to_soil}} for details.}

\item{keep_prev_soildepth}{A logical value.}

\item{keep_prev_soillayers}{A logical value.}

\item{verbose}{A logical value.}
}
\value{
A named list with elements: \describe{
  \item{updated}{
    A logical value indicating whether any layers/horizons
    have been added or removed.
  }
  \item{soil_layers}{An updated copy of \code{soil_layers}.}
  \item{soil_data}{An updated copy of \code{soil_data}.}
}
}
\description{
Add/remove soil layers to soil data
}
\examples{
soil_layers <- t(data.frame(
  site1 = c(20, 81, NA, NA, NA, NA),
  site2 = c(8, 20, 150, NA, NA, NA)
))

N <- ncol(soil_layers)
colnames(soil_layers) <- paste0("depth_L", seq_len(N))

soil_data <- cbind(
  sand = t(data.frame(
    site1 = c(0.5, 0.55, NA, NA, NA, NA),
    site2 = c(0.3, 0.35, 0.4, NA, NA, NA)
  )),
  TranspCoeff = t(data.frame(
    site1 = c(0.9, 0.1, NA, NA, NA, NA),
    site2 = c(0.7, 0.1, 0.2, NA, NA, NA)
  ))
)
colnames(soil_data) <- paste0(
  rep(c("sand", "TranspCoeff"), each = N),
  "_L",
  seq_len(N)
)

new_soils <- update_soil_profile(
  soil_layers = soil_layers,
  requested_soil_layers = c(5, 10, 20, 100, 200),
  soil_data = soil_data
)

}
