% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmosphere_calculate.R
\name{vp0}
\alias{vp0}
\title{Saturation vapor pressure}
\usage{
vp0(x)
}
\arguments{
\item{x}{A numeric vector of temperature \var{[C]}}
}
\value{
Saturation vapor pressure \var{[kPa]}
}
\description{
Saturation vapor pressure
}
\references{
Yoder, R. E., L. O. Odhiambo, and W. C. Wright. 2005. Effects of
  Vapor-Pressure Deficit and Net-Irradiance Calculation Methods on Accuracy
  of Standardized Penman-Monteith Equation in a Humid Climate Journal of
  Irrigation and Drainage Engineering 131:228-237.
}
