% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_soils.R
\name{check_texture_table}
\alias{check_texture_table}
\title{Check availability of soil texture values for a specified number of layers}
\usage{
check_texture_table(table_texture, n_layers, vars = c("sand", "clay"))
}
\arguments{
\item{table_texture}{A two-dimensional numeric object.
Rows represent sites and columns soil texture variables for each
soil layers \var{1, ..., n}. See examples.}

\item{n_layers}{A numeric vector. The number of soil layers at each site.}

\item{vars}{A vector of character strings. Soil texture variables to check
as base name of the column names of \code{table_texture}. See examples.}
}
\value{
A named list with \describe{
  \item{checks_passed}{A logical value. \code{TRUE} if no missing values.}
  \item{missing_N}{
    An integer n x p matrix for n sites and p variables (\code{vars}).
    The number of missing values for a site \var{i} and variable \var{k}
    among the specified soil layers (\code{n_layers[i]}).
  }
  \item{is_missing_anylayer}{
    A logical n x p matrix for n sites and p variables (\code{vars}).
    \code{TRUE} for a site \var{i} if at least one value of variable \var{k}
    is missing among the specified soil layer (\code{n_layers[i]}).
  }
  \item{is_missing_pctlayer}{\code{missing_N} divided by \code{n_layers}}
  \item{ids_sites_missing_anylayer}{
    An integer vector of site indices (row number in \code{table_texture})
    for which at least one variable in at least one layer is missing.
  }
  \item{ids_sites_missing_alllayers}{
    An integer vector of site indices (row number in \code{table_texture})
    for which at least one variable is missing in all layers.
  }
  \item{ids_sites_missing_somelayers}{
    \code{ids_sites_missing_anylayer} without
    \code{ids_sites_missing_alllayers}
  }
}
}
\description{
Check availability of soil texture values for a specified number of layers
}
\examples{
soils <- matrix(
  data = c(
    0.828, 0.963, NA, 0.065, 0.03, 0.03,
    0.57, NA, NA, 0.25, 0.03, 0.03
  ),
  nrow = 3,
  dimnames = list(NULL, paste0(c("sand_L", "clay_L"), rep(1:2, each = 2)))
)
N_horizons <- rep(2, 3)

texture_checks <- check_texture_table(
  table_texture = soils,
  n_layers = N_horizons,
  vars = c("sand", "clay")
)

# Does our soils table have any issues?
texture_checks[["checks_passed"]]

if (!texture_checks[["checks_passed"]]) {
  # How many sites have at least one missing value per variable?
  apply(texture_checks[["is_missing_anylayer"]], 2, sum)

  # Tabulate sites for number of layers with any missing sand values against
  # proportion of layers with any missing sand values
  addmargins(table(
    missing_pct = texture_checks[["is_missing_pctlayer"]][, "sand"],
    missing_N = texture_checks[["missing_N"]][, "sand"]
  ))

  # Tabulate sites number of layers with only missing values against
  # proportion of layers with only missing values
  ids <- texture_checks[["ids_sites_missing_alllayers"]]
  addmargins(table(
    missing_pct = texture_checks[["is_missing_pctlayer"]][ids, ],
    missing_N = texture_checks[["missing_N"]][ids, ]
  ))
}

}
