% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_simulations.R
\name{getStartYear}
\alias{getStartYear}
\title{Calculate start year of simulation period after \var{spin-up}}
\usage{
getStartYear(simstartyr, spinup_N = 1L)
}
\arguments{
\item{simstartyr}{An integer value. The calendar year when the \pkg{SOILWAT2}
was started.}

\item{spinup_N}{An integer value. The number of years that are considered
\var{spinup} for the simulation run.}
}
\description{
Calculate start year of simulation period after \var{spin-up}
}
